backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Task queue
celery[redis]>=5.4.0
redis>=5.1.0
kombu>=5.4.0

# Database — Supabase uses Postgres
psycopg2-binary>=2.9.10
supabase>=2.9.0
sqlalchemy>=2.0.35

# AI — Claude for lead qualification (Haiku)
anthropic>=0.39.0

# HTTP
httpx>=0.27.0
requests>=2.32.0

# Validation
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Utilities
python-dotenv>=1.0.1
phonenumbers>=8.13.0
